requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
colorama==0.4.4
DateTime==4.3
discord.py==1.6.0
distlib==0.3.1
filelock==3.0.12
idna==3.1
idna-ssl==1.1.0
importlib-metadata==1.7.0
importlib-resources==3.0.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
MouseInfo==0.1.3
multidict==5.1.0
Pillow==7.2.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pylint==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pytz==2020.5
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.0.26
wrapt==1.12.1
yarl==1.6.3
zipp==3.1.0
zope.interface==5.2.0
